requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
